+++ src/nvim/memline_defs.h
@@ -17,7 +17,7 @@ typedef struct info_pointer {
 
 typedef struct ml_chunksize {
   int mlcs_numlines;
-  long mlcs_totalsize;
+  int mlcs_totalsize;
 } chunksize_T;
 
 // Flags when calling ml_updatechunk()
